+++ openitgo-app/src/widgets/mod.rs
@@ -1,2 +1,3 @@
 pub mod progress_bar;
 pub mod thumbnail_progress_bar;
+pub mod volume_wedge;